pyproject.toml
[tool.poetry]
name = "grammar-checker"
version = "0.1.0"
description = ""
authors = ["Dhiraj Suvarna"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.31.0"
requests = "^2.31.0"
openai = "^1.12.0"
python-docx = "^1.1.0"
streamlit-feedback = "^0.1.3"
typer = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
